Fixed the fireworks, provided "undo" function
I realized that the celebratory fireworks pop-up dialog box wasn’t
appearing every time I won the game. So I fixed that.
Also, it was pretty frustrating that there was no way to “undo” the
action if you picked up a card from the show stack, but then didn’t
have a place to put it. So now, if that happens, you can put it back on
the show stack. Try it out!